Yukky markdown conflict with indented tags
When you indent some code on a block that is to be markdown rendered, it interprets the indents as
meaning that the code should be rendered as a literal code block. Substitute out leading spaces for lines
that start with spaces followed by a <. That should work. In my own head it does anyway.