BUGFIX: ordering of pages in tag listings
Tag listings were showing pages in order of when they were updated. This was a bit odd, as pages
might be updated subsequent to publication, meaning they could hop to the top of a listing. Now
pages are listed ordered by creation date.

TODO: A similar problem exists for writings